Refactor string input validators.
Pass the initial value of a validated input field (including string input
validated fields) to validated input field creators. Implement
create_input_field() overloads that take the result of input field
creators, create the input field with the validator and initial value
already set, and returnsa tuple of the input field and the validated
input field.

The type of the object must be explicitly passed to
create_string_validated_input_field_contents/set_string_validator,
and the parsed value gets passed to the closure in a std::optional instead
of using a pointer. This is more error-proof API, and losing the ability to
automatically deduce the type of the object is a small price to pay.