Automatically save and restore adjustable table column widths.
Fix missing call to layoutmanagers' uninstalling() method.

Table layout manager: replace adjustable_table_column_width bool, and name,
with a single object that defines whether the table columns are adjustable,
and if so what the table's name is. restore() gets removed. Table layout
manager's create() takes care of doing the restoral.

Store a reference to screen_positions in generic_window_handlers, it's no
longer necessary to pass in the WM_CLASS resource string, it can be
defaulted to the appid in screen_positions. Also, store a map
that tracks the labels of all widgets in the window.