Refactor widget and window configuration handling.
Remove duplicated code when saving and loading a widget or window
configuration.

Implement a config handle abstraction layer that gets constructed when
creating a widget (window code uses most of it, as is), which computes
the configuration's XPath, and creates it in the screen_positions XML
document if it doesn't exist. Widget type, namespace, and name, gets
specified only once, at this point.

Saving and loading the widget/window configuration uses the handle object
to obtain either a read or a write lock.