creator: improve input validation and error checking
Leaving the widget empty resulted in confusion, xpath_for() then
constructed an xpath for all elements, due to the unfortunate overload
of xpath_for being overloaded an an empty id was used to construct an
xpath for all elements. Disambiguate and use get_xpath_for_all(), for
that use case.

Provide an indication to each parameter widget whether it's required,
or not. This can be derived from the existing metadata. Change generators
for input widgets to check this parameter, and if it's required reject
empty input.

Have focusable_container return the editable_combobox_input_field().