Fix selection reset in lists.
When returning to a list the currently selected list item will be where
it was when keyboard or pointer focus left the list.

Remove current_keyed_element and the associated logic. This gets replaced
by a combined_focus flag that both keyboard and pointer focus update.
Check combined_focus when drawing list items, and do not reset the
current_element when losing focus, so we remember the current_element.

Fix home/end key logic to check individual items' enabled flag, and
move to the first or the last enabled item.